UMC fixes, fictive Nordsieks for required constants.  Generator PQ are calculated by equations again.

+++ DFW2/LoadFlow.cpp
@@ -1333,7 +1333,7 @@ void CLoadFlow::UpdatePQFromGenerators()
 					}
 
 					// вводим Q генератора в диапазон
-					pGen->Q = (std::max)((std::min)(pGen->Q, pGen->LFQmax * pGen->Kgen), pGen->LFQmin * pGen->Kgen);
+					pGen->Q = (std::max)((std::min)(static_cast<double>(pGen->Q), pGen->LFQmax * pGen->Kgen), pGen->LFQmin * pGen->Kgen);
 					pNode->Qg += pGen->Q;
 					// рассчитываем суммарные ограничения по генераторам
 					// в узлах остаются два ограничения - обычные (для суперузла - сумма)
